Taking a break!
OK so we have the ability to pull in the questions—next up is listening for a click on each button and sending back to server, where the server will see if it matches answer. Can use CSS to show which button you clicked rather than making the server do it.