(Kinda) fixed Next Question Button
I think breaking this whole thing out into smaller functions will help me understand what's going on. Next Question button works now but because the sign up button *also* asks for the next question, you end up getting two on screen.

In general it seems that to make this work multiplayer, we either need it to be synchronous or find a way to stop one player's questions impacting the other's...